backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
pillow==10.3.0
python-dotenv==1.0.1
#django-filter==24.2